Initialized the db variable by connecting to the Firebase database, refactored the userReducer to its own separate module, user, for better organization, created a createUser function using createAsyncThunk to asynchronously store new users in the Firebase database
