+++ store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import householdReducer from './householdReducer';
 import taskReducer from './taskReducer';
-import userReducer from './userReducer';
+import userReducer from './user/userReducer';
 
 export const store = configureStore({
   reducer: {
